backup-audit-k8s-app/requirements.txt
kubernetes==28.1.0
GitPython==3.1.40
pyyaml==6.0.1
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
pydantic==2.5.0
APScheduler==3.10.4 